[operator migration] Migrate merged momentum cpu/gpu kernels (#44300)
